requirements.txt
Django>=5.2
djangorestframework>=3.14
django-cors-headers>=4.3
django-filter>=23.3
drf-spectacular>=0.26
Pillow>=10.0
celery>=5.3
redis>=5.0
django-extensions>=3.2
python-decouple>=3.8
psycopg2-binary>=2.9
djangorestframework-simplejwt>=5.3
django-cleanup>=8.0
django-storages>=1.14
boto3>=1.28
python-dateutil>=2.8
openpyxl>=3.1
reportlab>=4.0
matplotlib>=3.8
seaborn>=0.12
pandas>=2.1
numpy>=1.24
